% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundataframe.R
\name{myfunction}
\alias{myfunction}
\title{creates a function to access data}
\usage{
myfunction(df, col)
}
\arguments{
\item{df}{is the dataframe}

\item{col}{is the column}
}
\description{
creates a function to access data
}
